Rework event domain
Created a separate DAO domain object for saving events due to changes needed to be made to datetime handling for Event objects created by data coming from the frontend. ISO 8601 strings in UTC cannot directly be converted into an Instant but have to use ZonedDateTime instead and ZonedDateTime cannot directly be saved to the database.